requirements.txt
Flask>=2.0
SQLAlchemy>=1.4
psycopg2-binary
pandas
numpy
scikit-learn
python-dateutil
flask-cors
python-dotenv
gunicorn
